src/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Login==0.6.3
Flask-WTF==1.2.1
Flask-Limiter==3.5.0
Flask-Cors==4.0.0
SQLAlchemy==2.0.25
Werkzeug==3.0.1
requests==2.31.0
python-dotenv==1.0.0
bcrypt==4.2.1
PyJWT==2.8.0
python-jose==3.3.0
numpy==1.26.3
pandas==2.1.4
scikit-learn==1.3.2
hmac==0.1.0
pycryptodome==3.19.0
redis==5.0.1
python-multipart==0.0.6
PyYAML==6.0.1
alembic==1.13.1
psycopg2-binary==2.9.9
python-dateutil==2.8.2
pytz==2024.2
six==1.16.0
itsdangerous==2.1.2
blinker==1.7.0
Jinja2==3.1.3
MarkupSafe==2.1.4
Werkzeug==3.0.1
pytest==7.4.4
pytest-cov==4.1.0
black==23.12.1
flake8==7.0.0
mypy==1.8.0
httpx==0.26.0
websockets==12.0
aiohttp==3.9.1
uvicorn==0.24.0
fastapi==0.109.0
python-multipart==0.0.6
pydantic==2.5.3
uvloop==0.19.0
alembic==1.13.1
psycopg2-binary==2.9.9
python-dateutil==2.8.2
pytz==2024.2
six==1.16.0
itsdangerous==2.1.2
blinker==1.7.0
Jinja2==3.1.3
MarkupSafe==2.1.4
Werkzeug==3.0.1
pytest==7.4.4
pytest-cov==4.1.0
black==23.12.1
flake8==7.0.0
mypy==1.8.0
httpx==0.26.0
websockets==12.0
aiohttp==3.9.1
uvicorn==0.24.0
fastapi==0.109.0
python-multipart==0.0.6
pydantic==2.5.3
uvloop==0.19.0
